+++ packages/es-markdown-core/src/visitor.ts
@@ -22,6 +22,7 @@ import type { Markup } from "./model/markup.js";
 import type { Paragraph } from "./model/paragraph.js";
 import type { SoftBreak } from "./model/soft-break.js";
 import type { Strikethrough } from "./model/strikethrough.js";
+import type { Mark } from "./model/mark.js";
 import type { Strong } from "./model/strong.js";
 import type { Table, TableCell, TableRow } from "./model/table.js";
 import type { Text } from "./model/text.js";
@@ -54,6 +55,7 @@ export interface Visitor<Result> {
     visitEmphasis(this: void, node: Emphasis): Result;
     visitStrong(this: void, node: Strong): Result;
     visitStrikethrough(this: void, node: Strikethrough): Result;
+    visitMark(this: void, node: Mark): Result;
     visitLink(this: void, node: Link): Result;
     visitImage(this: void, node: Image): Result;
     visitDirective(this: void, node: Directive): Result;
@@ -114,6 +116,8 @@ export function visit<Result>(node: Markup, visitor: Visitor<Result>): Result {
             return visitor.visitStrong(node);
         case "strikethrough":
             return visitor.visitStrikethrough(node);
+        case "mark":
+            return visitor.visitMark(node);
         case "link":
             return visitor.visitLink(node);
         case "image":