+++ packages/kotlin-markdown-core/src/commonMain/kotlin/com/nouprax/markdown/core/visitor/TreeDumper.kt
@@ -307,6 +307,10 @@ private class DumpVisitor(
         state.container("Strikethrough", node, children = node.content)
     }
 
+    override fun visitMark(node: Mark) {
+        state.container("Mark", node, children = node.content)
+    }
+
     override fun visitLink(node: Link) {
         state.container(
             "Link",