+++ packages/swift-markdown-core/Sources/MarkdownCore/Visitor/MarkupVisitor.swift
@@ -30,6 +30,7 @@ public protocol MarkupVisitor {
     mutating func visit(_ node: Emphasis) -> Result
     mutating func visit(_ node: Strong) -> Result
     mutating func visit(_ node: Strikethrough) -> Result
+    mutating func visit(_ node: Mark) -> Result
     mutating func visit(_ node: Link) -> Result
     mutating func visit(_ node: Image) -> Result
     mutating func visit(_ node: Directive) -> Result