+++ packages/swift-markdown-core/Sources/MarkdownCore/Visitor/MarkupWalkingVisitor.swift
@@ -46,6 +46,7 @@ public protocol MarkupWalkingVisitor {
     mutating func visit(_ node: Emphasis, phase: WalkPhase)
     mutating func visit(_ node: Strong, phase: WalkPhase)
     mutating func visit(_ node: Strikethrough, phase: WalkPhase)
+    mutating func visit(_ node: Mark, phase: WalkPhase)
     mutating func visit(_ node: Link, phase: WalkPhase)
     mutating func visit(_ node: Image, phase: WalkPhase)
     mutating func visit(_ node: Directive, phase: WalkPhase)
@@ -318,6 +319,13 @@ private struct WalkingDriver<WalkingVisitor: MarkupWalkingVisitor>: MarkupVisito
     }
 
     mutating func visit(_ node: Strikethrough) {
+        visitor.visit(node, phase: phase)
+        scheduleExit(node)
+        if phase == .entering {
+            for child in node.content.reversed() { actions.append(.enter(child)) }
+        }
+    }
+    mutating func visit(_ node: Mark) {
         visitor.visit(node, phase: phase)
         scheduleExit(node)
         if phase == .entering {